part2/pyproject.toml
[tool.poetry]
name = "skypro_vacancies"
version = "0.1.0"
description = ""
authors = ["Albina <aalmukhametova@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
psycopg2 = "^2.9.1"
Pillow = "^9.0.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
